pyproject.toml
[tool.poetry]
name = "addresses_service"
version = "0.1.0"
description = "Addresses Service"
authors = ["Arnoud Verrips <arnoudverrips@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
django-countries = "^7.3.2"
django-oauth-toolkit = "^2.0.0"
drf-spectacular = "^0.22.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
pre-commit = "^2.19.0"
coverage = "^6.4"
pytest-dotenv = "^0.5.2"

[tool.flake8]
max-line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
